+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.5
+
+* getter 'minTrailingIndex' was called on null(#18)
+
+## 2.0.4
+
+* keep the trailingChildren around when run out of children before reaching the scroll offset(#14)
+
 ## [2.0.2]
 
 * fix wrong layout when child's size is changed.